requirements.txt
# libraries
openai
python-telegram-bot
python-dotenv
jupyter
opencv-python